Allow duplicate ShaderOverride hashes
This allows multiple ShaderOverride sections to use the same hash on the
condition that every section specify 'allow_duplicate_hash = true'.

This is intended that scripts will set this flag on any sections they
create so that if a user creates a shaderoverride with the same hash
they will get a warning at first, but can choose to allow it so that
they can add their own commands without having to merge them with the
section from the script, allowing all the auto generated sections to be
grouped together.

The section names still have to be distinct, which offers protection
against scripts adding multiple identical sections if run multiple
times.

Note that this can mask the warning of duplicate non-command list
settings across different sections with the same hash, but that is part
of why it is required that each section acknowledge they are allowing
duplicates. We do at least try to avoid clobbering the values of these
settings with the defaults when parsing subsequent sections.