Fix regression getting the game resolution for texture hashes, etc.
The commit "GetResolutionFrom now gets the initial game Resolution"
introduced a regression by filling out the resolution from copy of the
swap chain description that won't have been made when upscaling is
disabled. Move the code that fills it out to the same location as the
code that records the original game resolution for upscaling.

Really, we should swap around the logic that uses 'GetResolutionFrom' -
record all resolutions, and use the heuristic only where we need it.