Hook up the assembly signature parsing code into 3DMigoto
This adds a "assemble_signature_comments" option to the d3dx.ini to turn
on the signature parser allowing shaders to be assembled entirely from
scratch and allowing their ISGN/OSGN/etc sections to be changed.

The reason for adding this as a tunable is twofold - firstly, although
the signature parser has been verified pretty extensively in
cmd_Decompiler against a range of test shaders and samples from games,
it has not been used live before and there is a real possibility we
might hit some variations we haven't been able to synthesize in the test
shaders to verify yet.

Secondly, while helping other shaderhackers I have spotted that
sometimes people write extra notes for themselves in these comments
blocks as to what input/output is what, and it is possible that these
type of modifications could cause the signature parser to fail.

The option is enabled in the template d3dx.ini for anyone starting a
fresh fix, but will be disabled by default if the option is missing for
backwards compatibility.