Assembler: Rename "debug.txt" to "3Dmigoto_disassembly_debug.txt"
"debug.txt" is a fairly generic name that could be used by some games
and doesn't really indicate that it is one of our files (I wasn't even
aware we generated it until just now). Rename it to
"3Dmigoto_disassembly_debug.txt" to indicate it is ours and add it to
uninstall.bat.