Add reset_per_frame_limits command
This command can be used on either a custom resource or shader to reset
the number of copies/executions that it considers it has performed so
far this frame. This is useful for games that render multiple scenes in
each frame (e.g. the cutscene at the start of Dreamfall Chapters Final
Cut Book 5 renders three scenes each frame - Zoe, News Anchors, and
finally the main scene), so that we can get fresh constant buffers and
run any shaders we need to process them for each new scene (if we can
find a way to detect the scene change), without having to run them for
every shader within a scene.