Add command lists on RTV and DSV clear calls
These command lists are similar to the existing [Present] command list
and allow us to execute arbitrary commands before and/or after clearing
a render or depth target.

Currently there is no way for the command list to know which RTV/DSV is
being cleared (we could add this relatively easily by passing it into
the command list like we do draw state, then adding a resource copy
target type for it, but let's leave that for later), but this is still
useful to trigger reset_per_frame_limits for games that render multiple
scenes in a frame.