Refactor deny_cpu_read and track_texture_updates Map/Unmap code
Both of these features can track and divert the buffer returned by a
Map() call and have their own separate code to achieve that, and we will
shortly have a third feature that can also do this. Refactor both of
these to use common code for the diversion, and a common area to store
information about the mapped resource.

We store the information about the mapping inside the context - because
Map() and Unmap() are called on a context they should not be called from
multiple threads at the same time, so storing it here means that we
shouldn't need to lock any global structures, which could help compared
to how the hash tracking used to store these.

This gets rid of the debug log for Map() since that information can be
determined from frame analysis, and removes the Texture2D specific code
from MapDenyCPURead since that could be used with any resource type now
that the common diversion code is generic enough to handle any type.

One gotcha that I had forgotten is that the hash tracking code will
track the hashes of any resource being written to, even if the map type
it is using is not one that we can divert (because writing to
D3D11_MAP_WRITE_NO_OVERWRITE blows up, and it seems the same might be
true of D3D11_MAP_WRITE). That makes me wonder if we should remove the
diversion altogether for the hash tracking code, but leaving it as is
for now.

track_texture_updates tested for regressions on Far Cry 4 (Open the
camera HUD with Z a couple of times usually exposes the texture hash
issue pretty quickly), and deny_cpu_read tested in Lichdom Battlemage
(look for disappearing geometry - even on the main menu).