Command List: Add "this" to get the depth/render target being cleared
The command lists triggered when clearing the depth or render target
currently have no way to find out which resource was being cleared. This
adds a new copy target "this" which will refer to the target being
cleared. This could also be used for other things in the future that are
triggered with a specific resource.