Hook up ShaderRegex engine
This hooks the new ShaderRegex engine up to the context to run just
before the draw/dispatch calls. While the engine currently does not have
anything that requires it to be run that late, some of the other
features I have in the pipeline do, and until everything is in and
working it is simplest to keep all the automatic shader patching in one
place. Later we can think about optimisations to patch shaders on load
time (especially for the non-hunting case).

The mReloadedShaders map gains two flags related to this:

- deferred_replacement_candidate indicates that a shader is a candidate
  for auto-patching. That is, it has not been loaded from ShaderFixes.
  This flag is updated as during shader (not config) reload to reflect
  the current state of the shader.

- deferred_replacement_processed indicates that the shader has been
  examined by the ShaderRegex engine since the last config reload,
  regardless of whether the engine decided to patch it or not. This
  prevents us spending time examining shaders that have already been
  checked.

If the ShaderRegex engine indicates that the shader was patched (and not
just matched) we log the resulting assembly, reassemble it back into
bytecode, create a shader through DirectX, update mReloadedShaders and
set the patched shader live just in time for the draw call.

This currently lacks any form of caching to disk - we can add this
later, but it is important to remember that we will also have to cache
the updated command lists, not just the patched shaders themselves.