Remove spaces before newlines in every log message globally
Some of our logging calls use spaces before newlines, and some don't,
and viewing a log file that has excess spaces at the end of some lines
is annoying, and while that is personal preference the fact that this is
not consistent just makes no sense, so strip excess whitespace globally.

If the motivation for separating out the newline escape sequence in the
code is to make it easier to spot, we should fix the more underlying
issue with using a logging function that doesn't automatically insert
newlines instead.

If you are looking for all LogInfo/LogDebug lines that don't include a
newline character, use this command:

    git grep -P 'Log(Info|Debug)W?\(L?"((?!\\n).)*$'

This commit is the result of the following sed command:

    sed -i 's/ \\n"/\\n"/' -- $(find . -name '*.h' -o -name '*.cpp')
    sed -i 's/ \\n\\n"/\\n\\n"/' -- $(find . -name '*.h' -o -name '*.cpp')