Fix newline errors in logging and tag calls that intentionally omit the newline
This command was used to locate all logging calls that lacked a newline
character on the same line:

    git grep -P 'Log(Info|Debug)W?\(L?"((?!\\n).)*$'

It then goes through and audits each result to see if the missing
newline was intentional or not. Where it does not appear to be
intentional it adds the missing newline, and where it does appear to be
intentional it appends "NoNL" to the log call.

For now the NoNL variants of the log calls are simple aliases to the
regular versions, but they are intended to indicate that the fact that
the newline is missing was not an accident, and can help if in the
future we switch the regular variants to something that inserts a
newline automatically.