Add command lists on UAV clears
Running command lists from RTV and DSV clears proved extremely useful to
detect scene changes, and doing the same for UAVs could also be handy
for similar purposes, or for when we are redirecting resources from the
game to our own, or want to grab the contents of a UAV before it is
cleared or modify it afterwards.

So while I don't have a pressing need for this right now, it is fairly
simple and could be useful in the future, so go ahead and hook up
command lists to the UAV clear calls.