Command Lists: Implement Resource/View clear operation
This implements a clear command, which can be used on any resource that
could be bound as a DSV, RTV or UAV (so, constant buffers and textures
are out of luck) by calling the appropriate DirectX clear API.

While we could already use a custom shader to clear a resource, this is
simpler and provides a hint to the driver that we don't need the data in
the resource - this is used in SLI to indicate that the driver does not
need to synchronise a buffer between cards for example, improving
performance (the SLI flags can also disable this synchronisation, but
they do so non-discriminatingly and therefore can cause problems if some
buffers did need to be synchronised).

If used on a view it will immediately try to clear that view, and if
used on a resource that doesn't have an associated view (which currently
includes custom resources, since their views are cached wherever they
are bound, not in the custom resource object itself) it will attempt to
create a view for that resource and clear it using that.

Examples:

Clear a custom resource (will attempt to create whatever type of view it
can based on the resource's bind flags):

    clear = ResourceFoo

Clear a currently bound UAV with 0.3 (float type clears only work with
certain UAV formats, otherwise it would fall back to an integer clear):

    clear = cs-u0 0.3

Clear a currently bound typed UAV with 0xffffffff with an integer clear:

    clear = cs-u0 int 0xffffffff

Clear currently bound o0 with red:

    clear = o0 1 0 0 1

Clear only the depth side of a depth/stencil view with the value 0.3:

    clear = depth oD 0.3

Clear the back buffer with orange:

    clear = bb 1 0.5 0 1

etc.

Beware that while testing this in Hellblade with SLI I discovered that
the driver seems to only clear certain types of resources in one eye,
including at least the back buffer and structured buffer UAVs. This is
not fully understood yet.