Add PCRE2 10.30
This is the latest version of the Perl Compatible Regular Expression
Library, extracted straight from upstream - it is not yet integrated
into our build system.

The intent is to use this to add support for dynamically patching
shaders on the fly. While C++11 does already include built in regular
expression support in std::regex, it lacks the named capture groups that
were first introduced in Python's re module and have since been adopted
by PCRE and any regular expression engine that expects to be taken
seriously.

And more to the point - I have a specific idea in mind of how this will
go in our d3dx.ini file that will require named capture groups to make
work.