Include static pcre2 libraries
To create these CMake was used to create 32bit and 64bit versions of the
pcre2 Visual Studio solution and project files, then building their
Release builds. The project files themselves aren't included, since
CMake put absolute paths in them (tsk tsk) and didn't have an option to
produce 32bit and 64bit targets in a single project file. With some
massaging we can probably include one, but for now we can make do with
the library and header files.

The config.h with the options I selected in CMake is included for
reference - in particular, the "8bit" (as in ASCII/UTF8) library is
enabled for both 32bit and 64bit (and renamed to add the platform bit
size), JIT is enabled because the performance gains it offers are pretty
compelling given that we need these to execute as quickly as possible to
minimise stuttering. Unicode is disabled since we shouldn't need it as
the shaders it will be working on should be restricted to ASCII.
Newlines are set to NL, because that seems to be what the shaders use.
The standard library is statically linked.