IniHandler: Limit the number of audible warning tones we emit for parse errors
There are certain types of errors that we can encounter while parsing
the d3dx.ini that can lead to a cascade of following warnings,
potentially playing error tones for quite some time with no way to stop
it, and if this takes longer than 30 seconds in a UE4 game it will
trigger UE4's watch dog timer killing the game.

One such common cascade failure is when a custom shader is missing or
fails to compile, as that causes the custom shader section to not be
added to our data structures and then every run command that refers to
it will issue a warning. Duplicate sections can also cause a cascade of
warnings, as every line in the duplicate is treated as being outside of
any section and will issue a warning.

This adds a wrapper around LogInfo and BeepFailure2 that will only issue
the audible warning for the first three issues encountered while loading
or reloading the d3dx.ini, but will always print out the warning message.

It also switches any warnings that were using LogInfoW to LogInfo with
%S for wide strings, because that means one less wrapper function.