Implement Shader Regex Engine
This implements a PCRE2 based regular expression engine that can match
patterns in assembly shaders and perform complex substitutions and
insertions to automatically patch shaders on the fly as they are used.

It supports allocating named temporary registers with automatic updating
of dcl_temps, has facilities to insert new declarations that ensure the
declaration has not already been inserted, and can add command lists to
any matching shader, even if no substitution was performed.

Usage examples will be included in a following commit that hooks these
up to sections in the d3dx.ini