IniHandler: Store raw lines in IniSectionVector
For most of our ini parsing we want to get key and value pairs with
extra whitespace around each stripped, because that is the norm for ini
files. But some of our sections are not parsed as pure ini files and
there are some cases where we would prefer to read the whole line
without any whitespace around the equals sign stripped, e.g. this will
be used by the upcoming shader regex engine, and could also be useful if
we want to match operators like == and != for conditional logic in the
command list in the future.

This will still strip whitespace at the beginning and end of the line as
a whole, and will still strip blank lines and comment lines.