Ini Handler: Add variant of GetIniString that fills out a std::string
The new Ini Handler was designed to very closely emulate the old one to
allow it to be swapped in with minimal impact on the rest of our code.
Doing this however bought along some baggage that we could do without -
in particular the old API converted the d3dx.ini to wide characters,
which has resulted in wide characters being propagated throughout much
of our code for no good reason (google wide characters considered
harmful for why we should aim to get rid of these). It also required us
to pre-allocate a sufficiently large buffer and would return the length
of the returned string - things we don't necessarily care about.

This adds a variant of GetIniString and GetIniStringAndLog that is a
starting point to get rid of some of that baggage. This returns a
std::string instead of filling out a wchar_t array, and eliminates the
need to pre-allocate a sufficient large buffer and just returns whether
the setting was found or not rather than the length.

Currently this converts the wide strings from our data structures back
to a regular strings, but as we eliminate more of the baggage of wide
characters we can eliminate the earlier conversion to wide strings and
store them as regular strings directly.