Command List: Add LinkCommandLists() call
This function will link two command lists together, so that executing
the destination list will execute the linked list. It works in much the
same way as the existing 'run = CommandList', but is intended to be used
to link command lists in dynamically at runtime. This will be used to
link the ShaderRegex command lists into the ShaderOverride command lists
of matching shaders.