Be less clever for compatibility with old toolchains
Seems the older toolchain doesn't like the = NULL initialiser here, but
I don't want these not initialised, so get rid of the union and go back
to casting.

What can I say - I was getting sick of casting ID3D11Resource pointers
into texture pointers everywhere and a union seemed to express the
reality that the pointer can be interpreted in multiple ways...

But that's no good if it needs a new toolchain, so go back to the more
traditional pointer casting.