Fix refcounting bug that could allow an original shader to be released
The auto patch engine exposed a refcounting bug elsewhere in the code,
where we could crash during RevertMissingShaders as the original shader
we had saved off had been released.

The sequence that could expose this is something like (in practice the
refcounts can be greater than stated here):

1.   Game calls CreateShader()
1.a) Shader is not found in ShaderFixes
1.b) Hunting or regex engine is enabled
1.c) 3DMigoto saves a copy of the original shader
1.d) Original shader returned to the game
1.e) Original shader refcount is at 1, but there are actually 2 references

2.   Game draws objects with the shader
2.a) Regex engine patches the shader
2.b) Regex engine stores the patched shader in the replace map
2.c) Shader is now a candidate for reverting to the original version
2.d) Original shader refcount is now at 1
2.e) Patched shader refcount is now at 2 (1 for creation, 1 for binding)

3.   Game unbinds the shader from the pipeline
3.a) Original shader refcount remains at 1 (wasn't bound)
3.a) Patched shader refcount is now at 1 (unbound)

3.   Game releases the shader
3.a) Original shader refcount drops to 0
3.b) Original shader is freed, pointers now invalid
3.c) Patched shader refcount remains at 1 (no way to know the
     original was released)

4.   User presses F10 to reload shaders
4.a) Shader is not found in ShaderFixes and flagged for reversion
4.b) 3DMigoto releases the patched shader
4.c) Patched shader refcount drops to 0 and is freed
4.d) 3DMigoto attempts to AddRef() on the original shader
4.e) --CRASH (or other undefined behaviour) due to invalid pointer dereference--

The solution is properly increase the refcount on the original shader
when we save the it off. This doesn't affect the code path where the
shader is replaced, since in that case the original shader that we save
is separate from the shader we return to the game, and they both have a
refcount of 1.

This also adds some comments to CreateShader since the different code
paths are subtly different and confusing.