Overlay: Show header lines from all shader types currently being hunted
Only showing one line can be annoying when hunting multiple shader types
simultaneously when lots of shaders have info lines, because once the
user has found a e.g. pixel shader the info line of the pixel shader is
masked by the vertex shader info line, and they have to continue cycling
the vertex shader to one that is not in ShaderFixes to reveal the pixel
shader info line.

I personally run into this all the time when working on a game fixed by
script and trying to identify a specific shader with marking_mode =
original, since I cycle all types simultaneously and a lot of the vertex
shaders will have tag lines thereby masking all other shader types.

Show all lines, but to keep things clean we only show lines and drop the
Y coordinate down if the shader actually has an info line to show. That
way, if the user is only hunting a single type of shader at a time they
will still only see one line (and as a reminder, the done_hunting key
is there to stop hunting all shader types, which will also clear these
lines).

Also, this now indicates which type of shader each line applies to since
that information was hidden before, and ensures that long lines do not
scroll off the left of the screen.