Ini Parsing: Another pass to remove manual logging
This goes through and switches almost all the remaining ini parsing call
sites to use the appropriate helper functions that provide automatic
logging, and removes the manual logging we have all over the place.

There's a new GetIniHash function for parsing the hash values in
ShaderOverride, TextureOverride and partner hashes, and a
GetIniHexString for the various mask fields in the render state
overrides.

Iteration has been left alone for now because it would need more work
and it's a feature we want to replace.