Overlay: Add an overlay_verbose option
This option shows the shader hashes in the overlay - since I very rarely
want to dump a shader just to find out which one I'm looking at
(especially when working with assembly or a fix entirely contained in
the d3dx.ini) I need to see this. It is disabled by default to keep the
interface clearer.