Fix long standing duplicate StereoParams bug
When reloading new shaders from ShaderFixes at runtime we would update
the bytecode in our replacement shader map with the newly compiled
bytecode, but that field is used for the original bytecode elsewhere in
3DMigoto, such as being used in CopyToFixes to decompile the shader and
store it in ShaderFixes (this would lead to amongst other things the
duplicate StrereoParams bug we sometimes see), or by the ShaderRegex
engine if the shader was later removed from ShaderFixes.

Remove the code that updated the bytecode and leave a comment in its
place.