Overlay: Fix unicode conversion and crash when rendering unsupported characters
There are two problems fixed by this commit:

Firstly, we were not correctly converting UTF8 into UTF16 when parsing
the shader's info line (we would just insert an extra zero byte after
each original byte), which this fixes using wstring_convert.

Secondly, if the font we are using lacks any characters we were trying
to display DirectTK would raise an exception. This call
SetDefaultCharacter to use ? for any unsupported characters.

This does not make any attempt to use a font with wider character
coverage.

As a side note - in this case it is borderline acceptable to store the
tag line as UTF16 because that's what the API expects it to be in, and
we don't otherwise do anything with it, but the simple truth is that the
API has been poorly designed as there is NEVER a good reason to use
UTF16 - it is literally the worst of all worlds and should be considered
harmful:

It is NOT fixed code point width like UTF32 (it's more like UTF8 in that
regards, but most people don't realise this) and requires special
handling like this to make work properly - just using wide characters
does not magically give the program unicode support. It bloats
everything, and is not an optimal size for CPU registers to work with.
There are only a few countries where text even ends up smaller in UTF16
than it does in UTF8, but that isn't even an advantage because any form
of compression offers better results again.

7-bit ASCII, UTF8 and UTF32 are the only sensible encodings (and UTF32
should only be used when it immediate access to characters by index is
required). Everything else is a mistake.

Windows has forced UTF16 on us, but that doesn't mean we should embrace
it.