Remove link-time dependency on nvapi.dll / nvapi64.dll
3DMigoto cannot be used on systems without an nvapi.dll because we have
it listed in the dependencies that the Windows dynamic linker must
satisfy in order to be able to load us:

    C:\cygwin64\home\dss\3DMigoto\x64\Zip Release>dumpbin /dependents d3d11.dll
    Microsoft (R) COFF/PE Dumper Version 14.00.24213.1
    Copyright (C) Microsoft Corporation.  All rights reserved.

    Dump of file d3d11.dll

    File Type: DLL

      Image has the following dependencies:

        nvapi64.dll
        XINPUT9_1_0.dll
        D3DCOMPILER_47.dll
        KERNEL32.dll
        USER32.dll
        GDI32.dll
        SHELL32.dll
        ole32.dll

This is not how nvapi.dll is supposed to work - we are supposed to
statically link against nvapi.lib / nvapi64.lib and call into it, and it
will attempt to load nvapi.dll / nvapi64.dll at run-time, failing
gracefully if it cannot.

The only reason that nvapi.dll ended up in our dependencies list, is
because we have a single call that bypasses the static library and goes
straight through to the dynamic library. This is our hacked
NvAPI_QueryInterface call that we use to notify our wrapper that the
following call is from us. With this one call removed, nvapi.dll drops
out of our dependencies list allowing 3DMigoto to operate on systems
that don't have a nvapi.dll.

This commit fixes the issue by loading nvapi.dll and locating the
NvAPI_QueryInterface address at runtime, in much the same way as the
static library does.

This has not been thoroughly tested on non-nvidia hardware to verify
that it works.