NVAPI Wrapper: Fail gracefully if could not load nvapi.dll
If the real nvapi.dll cannot be loaded we currently just
DoubleBeepExit() out, but this situation is not really so dire since
3DMigoto can still operate pretty well without nvapi, just with some
limited functionality. This is of particular interest to modding
communities outside of 3D Vision which would like 3DMigoto to work on
non-nvidia hardware.

Rather than dying, this changes our nvapi wrapper to fail gracefully so
that 3DMigoto can still run.

This has not been tested thoroughly, but at least it continues to work
if the real nvapi.dll is present, and forcing it to fail gives the same
failure state as trying to run the game when the real nvapi.dll is
missing.