Assembler: Fix crash in precision fixup of NANs on vs2015 branch
The disassembler was crashing when performing the precision fixup on NAN
values on the vs2015 branch. This turns out to be because d3dcompiler_47
uses symbolic values for NANs with fewer 0s than d3dcompiler_47, as in
 1.#INF0000 =  1.#INF00
-1.#INF0000 = -1.#INF00
-1.#IND0000 = -1.#IND00
 1.#QNAN000 =  1.#QNAN0
-1.#QNAN000 = -1.#QNAN0

These variants are missing from strToDWORD (in fact, +QNAN is missing
for d3dcompiler_46 output as well), which will cause the disassembler's
precision fixup code to process them, but without finding an exponent it
would dereferece a NULL pointer and crash.

The correct fix for these should go in strToDWORD, probably ignoring the
zeroes altogether, however I don't want to play with these symbolic
values without some test cases to verify all of the above cases, along
with other possible variants (differing bits in the significand and sign
fields) are handled correctly on both versions of d3dcompiler - some
simple test cases suggest that +/-IND may not be being disassembled
correctly.

For now, handle these in the precision fixup code by replacing them with
hex literals. These are guaranteed to assemble correctly and running
verification on a few test shaders are passing. The only real downside
is that l(0x7f800000) is less clear than l(1.#INF), but that will only
happen with d3dcompiler_47 or any cases that are missed by strToDWORD.