Command List: Always enforce no_view_cache for swap chain access
If we have a reference to the back buffer we will cause
IDXGISwapChain::ResizeBuffers() to fail with DXGI_ERROR_INVALID_CALL.
This can happen if a command list is used to bind it as a render target
as the render target view we create will hold a reference to the back
buffer. By default we will cache that view thereby holding a reference
on the back buffer preventing if from being resized.

The 'no_view_cache' copy option was introduced to solve this back in the
day, but exactly why that was necessary was not fully understood at the
time due to limited debuggability of games that exhibited the issue so
it was not made the default. UE4 gives us a much nicer error dialog
indicating which call failed, though linking the two was not obvious.

This patch introduces a 'forbid_view_cache' flag associated with the
back buffer and fake back buffer copy targets to pretty much universally
solve this. If the back buffer is still bound when ResizeBuffers() is
called it could still fail, but since the typical use case of binding
the back buffer is from a custom shader section, that section will take
care of unbinding it when it is complete.