Fix crashing if an ini setting overflows the destination buffer
I had used wcscpy_s, because VS insists we use the secure versions and
the documentation indicated it would return an error and set the
destination to an empty string on overflow... However, I missed the
little detail that it first calls an invalid parameter handler, the
behaviour of which will depend on the build configuration and whether a
random macro has been called to change it it - as it is, this handler
just killed us immediately if a setting was too long.

Change to use wcsncpy_s with _TRUNCATE, which will copy as much of the
string as will fit in the destination (but still guarantee it is NULL
terminated, unlike wcsncpy) and return an error on overflow. This fixes
the crash, and more closely emulates the behaviour of the original ini
parser, which would always truncate long settings.