Fix index buffer hashes and improve performance
This switches the index buffer hash to use mResources, which means that
it now works with texture filtering, checktextureoverride, presets, etc.

This commit tries to make the minimal changes necessary to make this
work, to enable us to compare the performance differences between the
two approaches of storing the buffer hash.

In a 40 second profile of Alien Isolation in game I see this:

GetPrivateData
--------------
Name                            | Inclusive Samples | Exclusive Samples | Inclusive Samples % | Exclusive Samples %
HackerContext::IASetIndexBuffer | 483               | 67                | 0.99                | 0.14
std::Hash<ID3D11Resource*...>   | 17                | 17                | 0.03                | 0.03

mResources
----------
Name                            | Inclusive Samples | Exclusive Samples | Inclusive Samples % | Exclusive Samples %
HackerContext::IASetIndexBuffer | 330               | 19                | 0.71                | 0.04
std::Hash<ID3D11Resource*...>   | 218               | 218               | 0.47                | 0.47

We need to compare inclusive samples of IASetIndexBuffer to count the
samples that are inside the GetPrivateData and hash calls, where we see
153 less samples in the same time period. We also see less exclusive
samples, which concerned me that we might have fluked being called less
in this run, but repeating the test a few times seems to provide more or
less consistent results.

We do see significantly more samples in std::Hash now, but that is to be
expected since it is now us performing the hash rather than DirectX, and
we do still see an overall reduction in the number of samples, so this
suggests that mResources is indeed faster than GetPrivateData (and we've
never even looked at optimising the hash function of our maps, which I
can pretty much guarantee is possible).

Further optimisations are possible, but the point of this commit is just
to compare the performance of GetPrivateData with an unordered map
lookup.