Restrict index buffer hash lookups to hunting=1 for performance
Since we removed the old indexbufferfilter there is no need to have
mCurrentIndexBuffer set unless we are hunting, so make all that logic
conditional on hunting to make this near zero cost in production.

Set mCurrentIndexBuffer to 0 before looking it up, to eliminate all the
else clauses.

Remove the LogDebugs - they are redundant since the FrameAnalysisLog
already records the same information.

Revert the way we print the resource hash in the frame analysis log to
the original code, to keep it consistent with every other function.