Overlay: Show selected index buffer hash on verbose overlay
Now that we can actually do something useful with the index buffers show
the hash of the currently selected index buffer in the verbose overlay
rather than having to get it from the log file. We don't show this in
the non-verbose overlay to keep things cleaner during typical use.