d3dx.ini: Document verbose_overlay option
I'm actually finding this option incredibly useful in Hellblade, since
it allows me to match up a shader to the log file or ShaderFixes without
dumping it, and for times when I already know the hash of the shader I'm
interested in and need to quickly and reliably cycle to it.

I actually don't think I could live without it now, and if I find this
useful it stands to reason that other people will as well, so document
it in the d3dx.ini. But, this option does risk that people pick up some
bad habbits (copying down the hash from the screen is error prone - I
don't know how many times I made a transcription error doing that in
Helix Mod), so leave it disabled by default and discourage its use in
the comment.