Fix visited vertex/pixel shader list when dumping index buffers
The list of visited vertex & pixel shaders printed in the log when
dumping an index buffer was bogus and contained far too many shaders,
which were all the shaders seen used with any index buffer that had been
selected since the last hunting reset key press.

Clear the list of visited shaders each time the selected index buffer is
cycled so that this list will only contain the shaders used with the
currently selected index buffer.