Overlay: Show currently selected render target hash when verbose
Since we're showing the rest of the selected objects while hunting we
might as well show the selected render target hash as well while
verbose. This one is stored as a handle rather than a hash, so we have
to look up the hash in the overlay.

We could switch the storage, but it would need more work, and render
target hashes tend to have a high probability of collisions so maybe
being able to select different render targets with the same hash might
be advantageous? I'm not positive - I always use frame analysis to find
these because it's pretty much impossible to tell what each render
target is responsible by appearance while hunting (we should maybe
think about changing these to show the render target rather than skip
it, which would be a lot more useful).