Clear visited render target list on done_hunting / timeout
The new info in the overlay showed that the list of total render targets
in the scene was growing endlessly as menus were opened and closed, and
was not being reset with the done_hunting key binding or on hunting
timeout.

Clear this list as well so it doesn't grow endlessly and to make hunting
render targets something that can be completed in a finite time.