Clear the visited index buffer/shader lists on done_hunting, not timeout
The comment indicated that these were suspect being cleared on hunting
timeout and now that I understand what they are for that is indeed
correct - they should be cleared only when cycling the index buffer or
shaders, which does not happen on hunting timeout. But, the done_hunting
key does cycle these to nothing, so it makes sense to clear these lists
at that time, so move the clear there.