Fix lost Present calls for Elex.
Adds QueryInterface override for HackerUnknown, because Elex calls to request itself for HackerDXGISwapChain and we otherwise are unhooked.

This just logs and returns This so that the game continues to get our HackerUnknown object, when the typeid's match.

Builds, but is untested.  Couldn't pull up a game I own that does a HackerUnknown::QueryInterface

Probably worth doing on HackerContext and HackerDevice as well, but I cannot be certain typeid won't crash until we can test it.