Fix performance regression in hash contamination checks with hunting enabled
Masterotaku noted that 1.2.66 performed significantly worse in Grim Dawn
when hunting was enabled (23-30fps down from 44-46fps), regardless of
whether hunting was set to 1 or 2 - only hunting=0 would recover this
performance loss.

This turns out to be due to the fact that setting orig_hash in
mResources for buffers would partially enable the hash contamination
detection, which does have a cost that can add up when a large number of
buffers are updated. This commit disables orig_hash for buffers to
eliminate this cost.

On the one hand, we may want hash contamination detection to indicate
when the hashes can't be relied on, same as we have for textures, and
this cost is only paid in hunting mode so we could look at it as being
not a huge concern, but - most people won't be using buffer hashes, so
it would be better not to pay this cost, and if this is any indication
of the performance cost of tracking buffer hash updates (and this is
nowhere near cost of actually enabling that since it is only detecting
contamination and is not re-hashing anything) then this shows that we
really don't want that enabled for buffers, and would need to find a
different solution if hash contamination ever proves to be a real
problem for buffers.

Reported-by: Masterotaku