Clean up HackerUnknown::QueryInterface.
Previous stage works for the test case, but is not correct, based on the documentation.

Change code to lookup the IUnknown interface for the 'this' object (mOrigUnknown), and the ppvObject that was returned.  In the case where they are identical, we can be sure it's OK to return 'this' instead of the ppvObject.

Added documentation here, and the reason for doing it this way.