Fix crash when issuing a draw call from a command list with hooking enabled
The command list code needs to perform a lot of DirectX calls that
bypasses 3DMigoto's processing, however when hooking is enabled the
GetOrigContext() / GetOrigDevice() calls would return the "real" DirectX
object, which was hooked and would therefore call back into 3DMigoto.

This caused some confusing frame analysis logs, where there seemed to be
some extra DirectX calls in odd places that didn't come from the game,
and in the worst case could case a crash if a command list triggered
from a draw call would itself issue a draw call, such as when trying to
run a custom shader section to override the render state. This would
then call back into the command list, which would call the draw call
again and so on until the stack ran out.

Most users of GetOrigContext() / GetOrigDevice() do legitimately need
the real DirectX object, whether it is hooked or not, but the command
list does need the trampoline object in this case. Add a
GetPassThroughOrigContext() / GetPassThroughOrigDevice() calls that will
return the trampoline object when hooked and the real object otherwise,
and switch the command list code to use them.